Refactor/db schema with grading and tos

+++ mobile/lib/data/datasources/local/assessments/impl/submission_datasource_mixin.dart
@@ -219,7 +219,7 @@ mixin SubmissionDataSourceMixin on AssessmentLocalDataSourceBase {
         submittedAt: rawRow['submitted_at'] != null ? DateTime.parse(rawRow['submitted_at'] as String) : null,
         autoScore: (rawRow['earned_points'] as num?)?.toDouble() ?? 0.0,
         finalScore: (rawRow['earned_points'] as num?)?.toDouble() ?? 0.0,
-        totalPoints: (rawRow['total_points'] as num?)?.toInt() ?? 0,
+        totalPoints: (rawRow['total_points'] as num?)?.toDouble() ?? 0.0,
         isSubmitted: rawRow['submitted_at'] != null,
       );
     } catch (e) {
@@ -290,7 +290,7 @@ mixin SubmissionDataSourceMixin on AssessmentLocalDataSourceBase {
         autoScore: (sub['earned_points'] as num?)?.toDouble() ?? 0.0,
         finalScore: (sub['earned_points'] as num?)?.toDouble() ?? 0.0,
         isSubmitted: sub['submitted_at'] != null,
-        totalPoints: (sub['total_points'] as num?)?.toInt() ?? 0,
+        totalPoints: (sub['total_points'] as num?)?.toDouble() ?? 0.0,
         answers: answers,
       );
     } catch (e) {
@@ -348,7 +348,7 @@ mixin SubmissionDataSourceMixin on AssessmentLocalDataSourceBase {
         submittedAt: row['submitted_at'] != null ? DateTime.parse(row['submitted_at'] as String) : null,
         autoScore: (row['earned_points'] as num?)?.toDouble() ?? 0.0,
         finalScore: (row['earned_points'] as num?)?.toDouble() ?? 0.0,
-        totalPoints: (row['total_points'] as num?)?.toInt() ?? 0,
+        totalPoints: (row['total_points'] as num?)?.toDouble() ?? 0.0,
         isSubmitted: row['submitted_at'] != null,
       )).toList();
     } catch (e) {
